initial refactoring prior to testing

+++ robot_dependencies_install.bash
@@ -5,7 +5,7 @@ set -e
 
 catkin_build()
 {
-  cd /home/"$USER"/catkin_ws
+  cd $CATKIN_DIR
   catkin build
 }
 
@@ -15,26 +15,6 @@ install_chrony_deps()
   sudo dpkg --configure -a
   sudo apt-get update
   sudo apt-get install gpsd gpsd-clients chrony
-}
-
-clone_ros_drivers()
-{
-  PROJECTS_DIR="/home/$USER/projects"
-
-  if [ ! -d "$PROJECTS_DIR" ]; then
-    mkdir $PROJECTS_DIR
-  fi
-
-  if [ -d $PROJECTS_DIR/ros_drivers ]; then
-    echo "ros_drivers already installed in $PROJECTS_DIR"
-    echo "ensure there is a symlink in catkin_ws"
-  else
-    cd $PROJECTS_DIR
-    echo "cloning ros_drivers to $PROJECTS_DIR..."
-    git clone --recursive git@github.com:BEAMRobotics/ros_drivers.git
-    echo "creating link in /home/"$USER"/catkin_ws/src/ to $PROJECTS_DIR"
-    ln -s $PROJECTS_DIR/ros_drivers /home/"$USER"/catkin_ws/src/
-  fi
 }
 
 install_ximea_deps()
@@ -63,31 +43,29 @@ install_ximea_deps()
 
 update_udev()
 {
-  # copy udev rules from inspector_gadget
+  # copy udev rules from ig_handle
   echo "copying udev rules..."
-  sudo cp ~/catkin_ws/src/ros_drivers/udev/* /etc/udev/rules.d/
+  sudo cp $CATKIN_DIR/src/ig_hangle/config/99-ig2_udev.rules /etc/udev/rules.d/
   sudo udevadm control --reload-rules && sudo service udev restart && sudo udevadm trigger
-}
-
-update_udev_ig2()
-{
-  # copy udev rules from inspector_gadget
-  echo "copying udev rules..."
-  sudo cp ~/catkin_ws/src/ig_hangle/config/99-ig2_udev.rules /etc/udev/rules.d/
-  sudo udevadm control --reload-rules && sudo service udev restart && sudo udevadm trigger
-  sudo cp ~/catkin_ws/src/ig_hangle/config/01-ig2_netplan.yaml /etc/netplan/
+  sudo cp $CATKIN_DIR/src/ig_hangle/config/01-ig2_netplan.yaml /etc/netplan/
 }
 
 install_ig_handle()
-{  
-  cd /home/"$USER"/catkin_ws/src/
+{
+  IG_HANDLE_DIR="ig_handle"
+  cd $CATKIN_DIR/src/
   echo "Installing ig-handle driver and dependencies..."
-  git clone git@github.com:BEAMRobotics/ig_handle.git   
+
+  if [ ! -d "$IG_HANDLE_DIR" ]; then
+    git clone git@github.com:BEAMRobotics/ig_handle.git
+  fi
+
   sudo apt-get install sharutils
-  
-  install_spinnaker_sdk  
+
+  install_spinnaker_sdk
   install_mti_sdk
   install_arduino_teensyduino
+  update_udev
 }
 
 install_gps()
@@ -122,76 +100,48 @@ enable_passwordless_sudo()
   echo "robot ALL=(ALL) NOPASSWD:ALL" | sudo tee -a /etc/sudoers
 }
 
-
 install_spinnaker_sdk()
 {
-  
-  echo "Installing spinnaker SDK..."
-  LB_DIR="spinnaker"
-  mkdir -p $DEPS_DIR
-  cd $DEPS_DIR
+  echo "Installing Spinnaker SDK..."
+  SP_DIR="spinnaker"
+  SP_SDK_DIR="spinnaker-2.4.0.143-amd64"
+  mkdir -p $DEPS_DIR && cd $DEPS_DIR
+  mkdir -p $SP_DIR && cd $SP_DIR
+
   sudo apt-get install libusb-1.0-0 libgtkmm-2.4-dev
   sudo apt-get install ros-$ROS_DISTRO-image-common
   sudo apt-get install ros-$ROS_DISTRO-image-exposure-msgs
   sudo apt-get install ros-$ROS_DISTRO-wfov-camera-msgs
-  sudo apt-get install ros-$ROS_DISTRO-image-proc  
+  sudo apt-get install ros-$ROS_DISTRO-image-proc
   # sudo apt-get install libavcodec57 libavformat57 libswscale4 libswresample2 libavutil55 
 
-  if [ ! -d "$LB_DIR" ]; then
-    echo "Don't have Spinnaker SDK Directory, creating & downloading SDK..."
-    mkdir -p $LB_DIR
-    cd $LB_DIR        
-    if [ "$ROS_DISTRO" = "kinetic" ]; then
-      # download Spinnaker SDK from sri_lab/robotics/software/apis/spinnaker-2.4.0.143-Ubuntu18.04-amd64-pkg.tar.gz
-      gdown --id 1se0fe_gu2IOxQHwVEdKLcOANbWdqoAAi 
-      tar fxv spinnaker-2.0.0.146-Ubuntu16.04-amd64-pkg.tar.gz
-      rm -rf spinnaker-2.0.0.146-Ubuntu16.04-amd64-pkg.tar.gz?dl=0      
-      cd spinnaker-2.0.0.146-amd64
-    elif [ "$ROS_DISTRO" = "melodic" ]; then
-      gdown --id 1_nT47nHHy6ugRxHH4frLV29wgCRhSRGF      
-      tar fxv spinnaker-2.4.0.143-Ubuntu18.04-amd64-pkg.tar.gz
-      rm -rf spinnaker-2.4.0.143-Ubuntu18.04-amd64-pkg.tar.gz      
-      cd spinnaker-2.4.0.143-amd64
-    fi
-    sudo sh install_spinnaker.sh
-    echo "Spinnaker SDK successfully installed."
-  else
-    echo "Already have spinnaker folder..."
-    cd $LB_DIR
-    if [ "$ROS_DISTRO" = "kinetic" ]; then      
-      gdown --id 1se0fe_gu2IOxQHwVEdKLcOANbWdqoAAi          
-      tar fxv spinnaker-2.0.0.146-Ubuntu16.04-amd64-pkg.tar.gz
-      rm -rf spinnaker-2.0.0.146-Ubuntu16.04-amd64-pkg.tar.gz
-      cd spinnaker-2.0.0.146-amd64
-    elif [ "$ROS_DISTRO" = "melodic" ]; then
-      gdown --id 1_nT47nHHy6ugRxHH4frLV29wgCRhSRGF
-      tar fxv spinnaker-2.4.0.143-Ubuntu18.04-amd64-pkg.tar.gz
-      rm -rf spinnaker-2.4.0.143-Ubuntu18.04-amd64-pkg.tar.gz
-      cd spinnaker-2.4.0.143-amd64
-    fi
-    sudo sh install_spinnaker.sh
-    echo "Spinnaker SDK successfully installed."
+  # download Spinnaker SDK from sri_lab/robotics/software/apis/spinnaker-2.4.0.143-Ubuntu18.04-amd64-pkg.tar.gz
+  if [ ! -d "$SP_SDK_DIR" ]; then
+    echo "Spinnaker SDK directory does not exist, downloading SDK..."
+    gdown 1_nT47nHHy6ugRxHH4frLV29wgCRhSRGF
+    tar fxv spinnaker-2.4.0.143-Ubuntu18.04-amd64-pkg.tar.gz
+    rm -rf spinnaker-2.4.0.143-Ubuntu18.04-amd64-pkg.tar.gz
   fi
+
+  cd $SP_SDK_DIR
+  sudo sh install_spinnaker.sh
+  echo "Spinnaker SDK successfully installed."
 }
 
 install_arduino_teensyduino()
 {
-  
   echo "Installing arduino and Teensyduino..."
-
-  echo "Installing rosserial..."
-  sudo apt-get install ros-$ROS_DISTRO-rosserial-arduino
   sudo apt-get install ros-$ROS_DISTRO-rosserial
-  echo "rosserial successfully installed."
+  sudo apt-get install ros-$ROS_DISTRO-rosserial-arduino
 
-  wget https://downloads.arduino.cc/arduino-1.8.13-linux64.tar.xz  
-  wget https://www.pjrc.com/teensy/td_153/TeensyduinoInstall.linux64  
-  wget https://www.pjrc.com/teensy/00-teensy.rules  
-  sudo cp 00-teensy.rules /etc/udev/rules.d/  
-  tar -xf arduino-1.8.13-linux64.tar.xz  
-  chmod 755 TeensyduinoInstall.linux64  
-  ./TeensyduinoInstall.linux64 --dir=arduino-1.8.13 
-   echo "arduino and Teensyduino successfully installed"
+  wget https://downloads.arduino.cc/arduino-1.8.13-linux64.tar.xz
+  wget https://www.pjrc.com/teensy/td_153/TeensyduinoInstall.linux64
+  wget https://www.pjrc.com/teensy/00-teensy.rules
+  sudo cp 00-teensy.rules /etc/udev/rules.d/
+  tar -xf arduino-1.8.13-linux64.tar.xz
+  chmod 755 TeensyduinoInstall.linux64
+  ./TeensyduinoInstall.linux64 --dir=arduino-1.8.13
+  echo "arduino and Teensyduino successfully installed."
 }
 
 install_virtual_box()
@@ -205,25 +155,31 @@ install_virtual_box()
   echo "Virtual Box successfully installed."
 }
 
-
 install_dt100()
 {
-  echo "Installing dt100 dependencies..."
+  echo "Installing DT100 driver and dependencies..."
+  DT100_DIR="dt100_driver"
+  cd $CATKIN_DIR/src/
+
+  if [ ! -d "$DT100_DIR" ]; then
+    echo "$DT100_DIR does not exist, cloning driver..."
+    git clone git@github.com:BEAMRobotics/dt100_driver.git
+  else
+    echo "$DT100_DIR exists."
+  fi
+
   install_virtual_box
-  # download Windows XP 32 DT100 virtual machine from sri_lab/robotics/software/vms/Windows_XP_32_DT100.ova
-  cd /home/$USER/catkin_ws/src
-  git clone git@github.com:BEAMRobotics/dt100_driver.git  
   VM_DIR="/home/$USER/virtual_machines/"
-  mkdir -p $VM_DIR
-  cd $VM_DIR
+  mkdir -p $VM_DIR && cd $VM_DIR
+
+  # download Windows XP 32 DT100 virtual machine from sri_lab/robotics/software/vms/Windows_XP_32_DT100.ova
   if [ ! -d "/home/$USER/VirtualBox\ VMs/Windows_XP_32_DT100" ]; then
     if [ ! -f "Windows_XP_32_DT100.ova" ]; then
-      echo "Importing virtual machine..."
-      gdown --id 1_X6_pstzYwIVQBICmkU4EMvxtBkayy_1
-      mv Windows_XP_32_DT100.ova Windows_XP_32_DT100.ova
+      echo "Importing virtual machine required by $DT100_DIR..."
+      gdown 1_X6_pstzYwIVQBICmkU4EMvxtBkayy_1
       vboxmanage import Windows_XP_32_DT100.ova
-    else 
-      echo "virtual machine has already been imported."
+    else
+      echo "virtual machine required by $DT100_DIR has already been imported."
     fi
   fi
   echo "dt100 driver and dependencies successfully installed."
@@ -233,26 +189,27 @@ install_dt100()
 install_mti_sdk()
 {
   echo "Installing MTI SDK..."
-  # creates a directory to store the MTI software
-  MT_DIR="MTIsoftware"  
-  mkdir -p $MT_DIR
+  MT_DIR="mti"
+  MT_SDK_DIR="MT_Software_Suite_linux-x64_2021.2"
+  mkdir -p $DEPS_DIR && cd $DEPS_DIR
+  mkdir -p $MT_DIR && cd $MT_DIR
 
-  cd $MT_DIR  
+  # download MTI SDK from sri_lab/robotics/software/apis/MT_Software_Suite_linux-x64_2021.2.tar.gz
+  if [ ! -d "$MT_SDK_DIR" ]; then
+    echo "MTI SDK directory does not exist, downloading SDK..."
+    gdown 1kTxxwwFHyDAJadEMhEjLIAN9_MnDgX-z
+    tar xvf MT_Software_Suite_linux-x64_2021.2.tar.gz
+    rm -rf MT_Software_suite_linux-x64_2021.2.tar.gz
+  fi
 
-  gdown --id 1kTxxwwFHyDAJadEMhEjLIAN9_MnDgX-z     
-  tar xvf MT_Software_Suite_linux-x64_2021.2.tar.gz 
-  rm -rf MT_Software_suite_linux-x64_2021.2.tar.gz 
-
-  cd MT_Software_Suite_linux-x64_2021.2 
-
+  cd $MT_SDK_DIR
   tar xvf mtmanager_linux-x64_2021.2.tar.gz
-  tar xvf magfieldmapper_linux-x64_2021.2.tar.gz  
+  tar xvf magfieldmapper_linux-x64_2021.2.tar.gz
   rm -rf mtmanager_linux-x64_2021.2.tar.gz
-  rm -rf magfieldmapper_linux-x64_2021.2.tar.gz 
-   
+  rm -rf magfieldmapper_linux-x64_2021.2.tar.gz
+
   chmod +x mtsdk_linux-x64_2021.2.sh
   bash mtsdk_linux-x64_2021.2.sh
 
   echo "MTI SDK sucessfully installed."
-    
 }
